BB-186: Move registration page to forms directory
